# docs/openapi-vendor-v1.yaml
openapi: "3.1.0"
info:
  title: "Paramita Vendor API Contract v1"
  version: "1.0.0"
  description: |
    Kontrak yang WAJIB diimplementasikan oleh setiap vendor yang terintegrasi dengan Paramita.
    Semua endpoint bersifat read-only (GET), kecuali token endpoint OAuth (POST).
    Vendor hanya mengembalikan data miliknya sendiri — Paramita yang menggabungkan.
    Lihat PRD §7–8 untuk aturan normatif lengkap.
servers:
  - url: "https://{vendor_base_url}"
    description: "URL dasar dikonfigurasi per vendor di panel Paramita"
    variables:
      vendor_base_url:
        default: "vendor.example.com"
security:
  - ApiKeyHeader: []
  - BearerToken: []
  - BasicAuth: []
  - OAuth2ClientCredentials: []
paths:
...
  /v1/inventory/summary:
    get:
      operationId: inventory.summary
      summary: "Ringkasan stok per scope"
      description: "Mengembalikan aggregate jumlah, stok, dan status stok untuk scope yang diminta."
      parameters:
        - $ref: "#/components/parameters/ut_code"
        - $ref: "#/components/parameters/program_codes"
        - $ref: "#/components/parameters/item_type"
        - $ref: "#/components/parameters/stock_status"
      responses:
        "200":
          description: "Ringkasan berhasil"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySummaryResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /v1/orders/summary:
    get:
      operationId: orders.summary
      summary: "Aggregate ringkasan DO"
      description: |
        Mengembalikan summary DO berdasarkan group_by=none/program/ut.
        group_by=vendor DILARANG dikirim ke vendor; Paramita membangun by_vendor sendiri.
      parameters:
        - name: group_by
          in: query
          required: true
          schema:
            type: string
            enum: ["none", "program", "ut"]
          description: "none=ungrouped, program=per prodi, ut=per unit daerah"
        - name: group_codes
          in: query
          style: form
          explode: false
          schema:
            type: array
            maxItems: 25
            uniqueItems: true
            items: { type: string }
          description: "Codes untuk group_by program/ut"
        - $ref: "#/components/parameters/ut_code"
        - $ref: "#/components/parameters/program_codes"
        - $ref: "#/components/parameters/process_status_code"
        - $ref: "#/components/parameters/ordered_from"
        - $ref: "#/components/parameters/ordered_to"
        - $ref: "#/components/parameters/period_code"
        - $ref: "#/components/parameters/search"
      responses:
        "200":
          description: "Summary berhasil"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /v1/orders/analytics:
    get:
      operationId: orders.analytics
      summary: "Agregat SLA, retry, dan distribusi untuk satu rentang waktu"
      parameters:
        - { name: ordered_from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: ordered_to, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: occurred_from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: occurred_to, in: query, required: true, schema: { type: string, format: date-time } }
        - $ref: "#/components/parameters/ut_code"
        - $ref: "#/components/parameters/program_codes"
      responses:
        "200":
          description: "Analytics berhasil"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalyticsResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API Key di header (rekomendasi default)"
    BearerToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      scheme: basic
    OAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: "https://vendor.example.com/oauth/token"
          scopes: {}
  parameters:
...
    search:
      name: search
      in: query
      schema:
        type: string
        minLength: 2
        maxLength: 100
      description: "Case-insensitive prefix search; min 2 char; empty dihilangkan"
    ut_code:
      name: ut_code
      in: query
      schema:
        type: string
        maxLength: 100
      description: "Kode unit daerah dari master (server-authoritative)"
    program_codes:
      name: program_codes
      in: query
      style: form
      explode: false
      schema:
        type: array
        maxItems: 20
        uniqueItems: true
        items: { type: string }
      description: "Sorted program codes, maks 20"
    item_type:
      name: item_type
      in: query
      schema:
        type: string
        enum: ["package", "book"]
    stock_status:
      name: stock_status
      in: query
      schema:
        type: string
        enum: ["shortage", "adequate", "surplus", "unknown"]
    process_status_code:
      name: process_status_code
      in: query
      schema:
        type: string
      description: "One or many comma-separated codes from 01..07"
    ordered_from:
      name: ordered_from
      in: query
      schema:
        type: string
        format: date-time
      description: "Inclusive start; RFC3339"
    ordered_to:
      name: ordered_to
      in: query
      schema:
        type: string
        format: date-time
      description: "Exclusive end; RFC3339"
...
    period_code:
      name: period_code
      in: query
      schema:
        type: string
      description: "Academic period code; optional"
...
  schemas:
...
    ObjectMeta:
      type: object
      required: [generated_at, data_as_of, scope]
      properties:
        generated_at:   { type: string, format: date-time }
        data_as_of:     { type: string, format: date-time, description: "Waktu maksimum data bisnis yang membentuk response" }
        scope:
          type: object
          required: [ut_code, program_codes]
          properties:
            ut_code:       { type: ["string", "null"] }
            program_codes: { type: array, items: { type: string } }
...
    InventorySummaryRow:
      type: object
      required: [item_type, stock_quantity, record_count, shortage_count, adequate_count, surplus_count, unknown_count]
      properties:
        item_type:      { type: string, enum: [package, book] }
        stock_quantity: { type: integer, minimum: 0 }
        record_count:   { type: integer, minimum: 0 }
        shortage_count: { type: integer, minimum: 0 }
        adequate_count: { type: integer, minimum: 0 }
        surplus_count:  { type: integer, minimum: 0 }
        unknown_count:  { type: integer, minimum: 0 }
...
    SummaryRow:
      type: object
      required: [total_orders, status_counts, status_code_counts, sla_counts, completed_sla_seconds_sum, completed_sla_sample_count]
      properties:
        total_orders:             { type: integer, minimum: 0 }
        status_counts:
          type: object
          required: [on_process, on_delivery, retry, returned, delivered]
          properties:
            on_process:  { type: integer, minimum: 0 }
            on_delivery: { type: integer, minimum: 0 }
            retry:       { type: integer, minimum: 0 }
            returned:    { type: integer, minimum: 0 }
            delivered:   { type: integer, minimum: 0 }
        status_code_counts:
          type: object
          required: ["01","02","03","04","05","06","07"]
          properties:
            "01": { type: integer, minimum: 0 }
            "02": { type: integer, minimum: 0 }
            "03": { type: integer, minimum: 0 }
            "04": { type: integer, minimum: 0 }
            "05": { type: integer, minimum: 0 }
            "06": { type: integer, minimum: 0 }
            "07": { type: integer, minimum: 0 }
        sla_counts:
          type: object
          required: [on_sla, over_sla, not_applicable, unknown]
          properties:
            on_sla:         { type: integer, minimum: 0 }
            over_sla:       { type: integer, minimum: 0 }
            not_applicable: { type: integer, minimum: 0 }
            unknown:        { type: integer, minimum: 0 }
        completed_sla_seconds_sum:  { type: integer, minimum: 0 }
        completed_sla_sample_count: { type: integer, minimum: 0 }
    GroupRow:
      type: object
      required: [group_code, summary]
      properties:
        group_code: { type: string, maxLength: 100 }
        summary:    { $ref: "#/components/schemas/SummaryRow" }
...
    InventorySummaryResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/InventorySummaryRow" }
        meta: { $ref: "#/components/schemas/ObjectMeta" }
...
    OrderSummaryResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          oneOf:
            - $ref: "#/components/schemas/SummaryRow"
            - type: array
              items: { $ref: "#/components/schemas/GroupRow" }
        meta: { $ref: "#/components/schemas/ObjectMeta" }
...
    AnalyticsResponse:
      type: object
      required: [data, meta]
      description: "Shape rinci setiap row mengikuti template aktif orders.analytics yang ditampilkan pada portal vendor. Template database adalah validator runtime utama."
      properties:
        data:
          type: object
          required: [sla, sla_details, retries, retry_details, distribution]
          properties:
            sla: { type: array, items: { type: object } }
            sla_details: { type: array, items: { type: object } }
            retries: { type: array, items: { type: object } }
            retry_details: { type: array, items: { type: object } }
            distribution: { type: array, items: { type: object } }
        meta:
          allOf:
            - $ref: "#/components/schemas/ObjectMeta"
            - type: object
              required: [ordered_from, ordered_to, occurred_from, occurred_to]
              properties:
                ordered_from: { type: string, format: date-time }
                ordered_to: { type: string, format: date-time }
                occurred_from: { type: string, format: date-time }
                occurred_to: { type: string, format: date-time }
    ProblemResponse:
      type: object
      required: [type, title, status, detail, instance, request_id]
      properties:
        type:       { type: string, format: uri }
        title:      { type: string, maxLength: 255 }
        status:     { type: integer, minimum: 100, maximum: 599 }
        detail:     { type: string, maxLength: 1000 }
        instance:   { type: string, maxLength: 2048 }
        request_id: { type: string, maxLength: 100 }
  responses:
    BadRequest:
      description: "400 Bad Request"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemResponse" }
    Unauthorized:
      description: "401 Unauthorized — belum login"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemResponse" }
    Forbidden:
      description: "403 Forbidden — di luar scope"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemResponse" }
...
    UnprocessableEntity:
      description: "422 Unprocessable Entity — parameter tidak valid"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemResponse" }
    TooManyRequests:
      description: "429 Too Many Requests"
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemResponse" }
...